Entity와 DBMS간 컬럼명 통일작업 및 수정작업
탐방 내에서 문화재 리스트 불러오는 기능 아직 미구현

+++ BackEnd/dmobile/src/main/java/com/ssafy/dmobile/relic/repository/ListDataRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.dmobile.relic.repository;
 import com.ssafy.dmobile.relic.entity.ListData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ListDataRepository extends JpaRepository<ListData, Integer> {
+public interface ListDataRepository extends JpaRepository<ListData, Long> {
 }
